commands: move diff code to diff.rs
